refactor: StudyRoom, StudyRoomHeader 파일명 수정
