+++ client/src/App.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter, RouterProvider, Navigate } from 'react-router-dom'
 import ProtectedRoute from '@components/common/ProtectedRoute';
 import Home from '@components/Home/Home';
 import StudyRoomList from '@components/StudyRoomList/index';
-import StudyRoom from '@components/StudyRoom/StudyRoom';
+import StudyRoom from '@components/StudyRoom/index';
 
 const router = createBrowserRouter([
   {
