+++ client/src/components/Home/Home.tsx
@@ -17,20 +17,17 @@ const Home = () => {
   return (
     <div className="flex h-screen w-screen items-center justify-center">
       <div className="flex h-[32rem] w-[68.5rem] items-center justify-between">
-        <div className="flex h-[32rem] w-[32rem] flex-col justify-center gap-9">
+        <div className="flex h-[32rem] w-[32rem] flex-col justify-center gap-12">
           <div className="flex flex-col items-center">
             <h3 className="text-3xl font-semibold">실시간 학습 기록 플랫폼</h3>
             <h1 className="text-9xl font-bold">GOMZ</h1>
           </div>
-          <div className="flex flex-col items-center gap-5">
-            <button className="font- bg-gomz-black h-[5rem] w-[20rem] rounded-full text-3xl font-semibold text-white">
-              로그인
-            </button>
+          <div className="flex flex-col items-center">
             <button
               onClick={handleGuestLogin}
-              className="border-gomz-black h-[5rem] w-[20rem] rounded-full border bg-white text-3xl font-semibold"
+              className="bg-gomz-black h-[5rem] w-[20rem] rounded-full text-3xl font-semibold text-white transition-transform hover:scale-105"
             >
-              로그인없이 이용하기
+              시작하기
             </button>
           </div>
         </div>
