+++ client/src/components/StudyRoom/ControlBar.tsx
@@ -42,7 +42,7 @@ const ControlBar = ({
     <div className={`relative flex gap-10 ${className}`}>
       <div className="bg-gomz-gray-300 flex h-10 w-20 items-center rounded-full">
         <button
-          className="bg-gomz-black flex h-10 w-12 items-center justify-center rounded-full"
+          className="bg-gomz-black flex h-10 w-12 items-center justify-center rounded-full transition-transform hover:scale-105"
           onClick={() => toggleVideo().then((result) => setIsVideoOn(result))}
         >
           <Icon
@@ -59,7 +59,7 @@ const ControlBar = ({
       </div>
       <div className="bg-gomz-gray-300 flex h-10 w-20 items-center rounded-full">
         <button
-          className="bg-gomz-black flex h-10 w-12 items-center justify-center rounded-full"
+          className="bg-gomz-black flex h-10 w-12 items-center justify-center rounded-full transition-transform hover:scale-105"
           onClick={() => setIsMicOn(toggleMic())}
         >
           <Icon id={isMicOn ? 'mic' : 'mic-off'} className="text-gomz-white h-6 w-6 fill-current" />
@@ -73,7 +73,7 @@ const ControlBar = ({
       </div>
       <button
         onClick={toggleChat}
-        className="bg-gomz-black relative flex h-10 w-20 items-center rounded-full"
+        className="bg-gomz-black relative flex h-10 w-20 items-center rounded-full transition-transform hover:scale-105"
       >
         <div className="flex h-10 w-12 items-center justify-center rounded-full">
           <Icon
@@ -87,7 +87,7 @@ const ControlBar = ({
       </button>
       <button
         onClick={exitRoom}
-        className="bg-gomz-red flex h-10 w-20 items-center justify-center rounded-full"
+        className="bg-gomz-red flex h-10 w-20 items-center justify-center rounded-full transition-transform hover:scale-105"
       >
         <Icon id="call-end" className="text-gomz-white h-7 w-7 fill-current" />
       </button>
